+++ flickr-shortcode-importer.php
@@ -132,7 +132,6 @@ EOD;
 
 		$links = array(
 			'<a href="http://aihr.us/about-aihrus/donate/"><img src="https://www.paypalobjects.com/en_US/i/btn/btn_donate_SM.gif" border="0" alt="PayPal - The safer, easier way to pay online!" /></a>',
-			'<a href="http://aihr.us/downloads/flickr-shortcode-importer-premium-wordpress-plugin/">Purchase Flickr Shortcode Importer Premium</a>',
 		);
 
 		$input = array_merge( $input, $links );